fix(core): split into several files
